INS-568 remove comments and useless files
